+++ src/Sharpbrake.Http.Module/AspNetHttpContext.cs
@@ -24,6 +24,8 @@ namespace Sharpbrake.Http.Module
         public string Action { get; set; }
         public string Component { get; set; }
 
+        public string Severity { get; set; }
+
         public AspNetHttpContext(HttpContext context)
         {
             var request = context.Request;