+++ test/Sharpbrake.Client.Tests/Mocks/FakeHttpContext.cs
@@ -17,5 +17,6 @@ namespace Sharpbrake.Client.Tests.Mocks
         public string UserName { get; set; }
         public string Action { get; set; }
         public string Component { get; set; }
+        public string Severity { get; set; }
     }
 }